Remove deprecated code.
